Refactored the service to include the directory parameters. Modifed the FileSystemSnapshot class to save snapshot in directory being snapshot. In addition add capability to filter out the file from snapshot results.

+++ SyncSharkEngine/FileSystem/FileInfoFacade.cs
@@ -47,7 +47,8 @@ namespace SyncSharkEngine.FileSystem
 
         public void Delete()
         {
-            m_FileInfo.Delete();
+            if (m_FileInfo.Exists)
+                m_FileInfo.Delete();
         }
     }
 }
